+++ apps/web/src/app/App.tsx
@@ -1,7 +1,7 @@
-import { Stack } from '@/shared/lib/stackflow'
+import { Stack } from '@/app/routes/stackflow'
 import '@stackflow/plugin-basic-ui/index.css'
 import QueryProvider from '@/app/Provider/QueryProvider'
-import { SessionProvider } from '@/shared/contexts/sessionContext'
+import { SessionProvider } from '@/app/Provider/contexts/sessionContext'
 import { Toaster } from '@/shared/ui/common/Sonner'
 import { OverlayProvider } from '@/shared/hooks/useOverlay'
