+++ apps/web/src/features/image-management/hooks/index.ts
