+++ apps/web/src/features/image-management/hooks/useImageUpload.ts
