+++ apps/web/src/pages/auth/ui/NickNamePage.tsx
@@ -11,8 +11,8 @@ import { type ExtendedKyHttpError } from '@/types/api/common'
 import { AppLayout } from '@/shared/ui/layout/AppLayout'
 import { BackArrow } from '@/assets/BackArrow'
 import { authGet } from '@/entities/auth/api'
-import { useFlow } from '@/shared/lib/stackflow'
-import { useSessionContext } from '@/shared/contexts/sessionContext'
+import { useFlow } from '@/app/routes/stackflow'
+import { useSessionContext } from '@/app/Provider/contexts/sessionContext'
 import WhiteBallogLogo from '@/assets/whiteBallogLogo.svg?react'
 import { useStack } from '@/shared/hooks/stackflow/useStack'
