+++ apps/web/src/pages/auth/ui/TeamSelectPage.tsx
@@ -1,7 +1,7 @@
 import { AppScreen } from '@stackflow/plugin-basic-ui'
 
 import { TeamSelectionForm } from '@/features/auth/ui'
-import { useFlow } from '@/shared/lib/stackflow'
+import { useFlow } from '@/app/routes/stackflow'
 import type { TeamKey } from '@/shared/constants/teams'
 import { AppLayout } from '@/shared/ui/layout/AppLayout'
 import { BackArrow } from '@/assets/BackArrow'
