+++ apps/web/src/pages/auth/ui/TermAgreePage.tsx
@@ -2,8 +2,8 @@ import { AppScreen } from '@stackflow/plugin-basic-ui'
 
 import { BackArrow } from '@/assets/BackArrow'
 import { AppLayout } from '@/shared/ui/layout/AppLayout'
-import { TermAgreementForm } from '@/features/auth/ui/TermAgreementForm'
-import { useFlow } from '@/shared/lib/stackflow'
+import { TermAgreementForm } from '@/pages/auth/ui/TermAgreementForm'
+import { useFlow } from '@/app/routes/stackflow'
 import { type TermItemData } from '@/entities/term/model/term.type'
 import { TERM_ID } from '@/entities/term/constants/termId'
 import WhiteBallogLogo from '@/assets/whiteBallogLogo.svg?react'
