+++ apps/web/src/pages/live-recording/ui/EmotionVoteWidget.tsx
@@ -6,8 +6,8 @@ import { EmotionButton } from '@/shared/ui/common'
 import { useEmotionVote } from '@/pages/live-recording/contexts/EmotionVoteContext'
 import type { EmotionType } from '@/entities/record/model/emotion.type'
 
-import { getScale } from './utils/getScale'
-import { getGridRatio } from './utils/getGridRatio'
+import { getScale } from '../utils/getScale'
+import { getGridRatio } from '../utils/getGridRatio'
 
 interface EmotionVoteWidgetProps extends ComponentProps<'div'> {
   emotions?: EmotionType
