+++ apps/web/src/pages/live-recording/utils/calculateGradientColor.ts
