+++ apps/web/src/pages/live-recording/utils/getGridRatio.ts
