+++ apps/web/src/pages/live-recording/utils/getScale.ts
