+++ apps/web/src/pages/mypage/ui/ChangeTeamSelectPage.tsx
@@ -2,12 +2,12 @@ import { AppScreen } from '@stackflow/plugin-basic-ui'
 import { toast } from 'sonner'
 
 import { TeamSelectionForm } from '@/features/auth/ui'
-import { useFlow } from '@/shared/lib/stackflow'
+import { useFlow } from '@/app/routes/stackflow'
 import { AppLayout } from '@/shared/ui/layout/AppLayout'
 import { BackArrow } from '@/assets/BackArrow'
 import { TEAMS, type TeamKey } from '@/shared/constants/teams'
-import { useUpdateMyTeamMutation } from '@/shared/hooks/auth/useUpdateMyInfoMutation'
-import { useSessionContext } from '@/shared/contexts/sessionContext'
+import { useUpdateMyTeamMutation } from '@/entities/auth/hooks/useUpdateMyInfoMutation'
+import { useSessionContext } from '@/app/Provider/contexts/sessionContext'
 import WhiteBallogLogo from '@/assets/whiteBallogLogo.svg?react'
 
 const isSignUpFlow = false
@@ -25,7 +25,7 @@ const ChangeTeamSelectPage = () => {
       {
         onSuccess: () => {
           pop()
-          toast("응원 팀 변경이 완료되었습니다!")
+          toast('응원 팀 변경이 완료되었습니다!')
           setUser({ ...user, baseballTeam: selectedTeam })
         },
       },
