+++ apps/web/src/widgets/navigation/ui/GlobalNavigationBar.tsx
@@ -3,7 +3,7 @@ import { useState, useEffect, type ComponentType } from 'react'
 import HomeGNB from '@/assets/HomeGNB'
 import RecordGNB from '@/assets/RecordGNB'
 import MypageGNB from '@/assets/MyPageGNB'
-import { useFlow, activities, actions } from '@/shared/lib/stackflow'
+import { useFlow, activities, actions } from '@/app/routes/stackflow'
 import { cn } from '@/shared/lib/classnames'
 import { GNBButton } from '@/shared/ui/common'
